Bug 1298010 - shut down worker type if deploymentId has updated in worker type config

+++ clean-builds.sh
@@ -18,7 +18,13 @@ rm -rf target
 mkdir target
 
 export GO_DOWNLOAD_DIR="$(mktemp -d -t go_download.XXXXXXXXXX)"
-curl -o target/go.tar.gz -L https://storage.googleapis.com/golang/go1.7.3.darwin-amd64.tar.gz
+
+if [ -f ~/go.tar.gz ] && md5 ~/go.tar.gz | grep -q d69f55f3174d3ee74c9bf7feb917d55f; then
+    cp ~/go.tar.gz target/go.tar.gz
+else
+  curl -o target/go.tar.gz -L https://storage.googleapis.com/golang/go1.7.3.darwin-amd64.tar.gz
+fi
+
 tar -C "${GO_DOWNLOAD_DIR}" -xf target/go.tar.gz
 export GOROOT="${GO_DOWNLOAD_DIR}/go"
 export PATH="${GOROOT}/bin:${PATH}"
